reorganize react-dom internals to match react
