+++ packages/react-dom/index.classic.fb.js
@@ -9,17 +9,16 @@
 
 import {isEnabled} from './src/events/ReactDOMEventListener';
 
-import {__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED} from './src/client/ReactDOM';
+import Internals from './src/ReactDOMInternals';
 
 // For classic WWW builds, include a few internals that are already in use.
-Object.assign((__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED: any), {
+Object.assign((Internals: any), {
   ReactBrowserEventEmitter: {
     isEnabled,
   },
 });
 
 export {
-  __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,
   createPortal,
   createRoot,
   hydrateRoot,
@@ -36,3 +35,5 @@ export {
   unstable_runWithPriority, // DO NOT USE: Temporarily exposed to migrate off of Scheduler.runWithPriority.
   version,
 } from './src/client/ReactDOM';
+
+export {Internals as __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED};
