+++ packages/react-dom/index.js
@@ -9,8 +9,8 @@
 
 // Export all exports so that they're available in tests.
 // We can't use export * from in Flow for some reason.
+export {default as __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED} from './src/ReactDOMInternals';
 export {
-  __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,
   createPortal,
   createRoot,
   hydrateRoot,
