+++ packages/react-dom/index.stable.js
@@ -7,8 +7,8 @@
  * @flow
  */
 
+export {default as __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED} from './src/ReactDOMInternals';
 export {
-  __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,
   createPortal,
   createRoot,
   hydrateRoot,
